Fixed issues with errors, vector of hands works
Before: multiple errors because player.hand was not a vector.  Then I
changed to vector and got numerous errors; obviously.  This update
removed them. Some testing was done, although further testing is
required.  In the meantime, the program works as it did be for the hand
to vector<hand> change.